restructure the api code
